fix(nextly): mark enforced content reads dynamic so they escape the route cache
